+++ src/main/kotlin/io/github/trcdevelopers/clayium/common/blocks/ItemBlockClayLaserReflector.kt
@@ -1,7 +1,7 @@
 package io.github.trcdevelopers.clayium.common.blocks
 
 import io.github.trcdevelopers.clayium.api.util.CUtils
-import io.github.trcdevelopers.clayium.client.renderer.LaserReflectorItemStackRenderer
+import io.github.trcdevelopers.clayium.client.renderer.item.LaserReflectorItemStackRenderer
 import net.minecraft.item.ItemBlock
 
 class ItemBlockClayLaserReflector(block: BlockClayLaserReflector) : ItemBlock(block) {
